token_server/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
livekit-api==1.0.4
flask-cors==4.0.0
